rocks-command: discarege caches before start VM
when we start a vm which uses most of the RAM sometime it fails with

Unable to read from monitor: Connection reset by peer

If you look in /vat/log/libvirt/qemu/qemu.log there is an error
which indicates that there is not enough RAM memory.

To avoid this you need to drop caches (which does not make any
sense).